Update - Deprecate old event listeners file
